Modified player and main camera
Added movement script to player and added made main camera a child to player. Added background audio to camera along with two image effects components. I've also added a movement script to the player and unchecked the mesh render component in order to not cast a shadow or show the players body.